split runtests.jl to separate files
