+++ src/utils.jl
@@ -46,8 +46,8 @@ end
 import Base: one, conj
 import AbstractAlgebra: degree
 
-Base.one(G::Generic.SymmetricGroup{I}) where I = Perm(I(1):G.n)
-Base.one(g::Perm{I}) where I = Perm(I(1):degree(g))
+Base.one(G::Generic.SymmetricGroup{I}) where I = Perm(G.n)
+Base.one(g::Perm{I}) where I = Perm(degree(g))
 
 AbstractAlgebra.degree(p::Perm{I}) where I = I(length(p.d))
 
@@ -73,3 +73,11 @@ end
 
 Base.conj(h::GroupElem, g::GroupElem) = conj!(h, h, g)
 Base.:(^)(h::Perm, g::Perm) = conj(h,g)
+
+function AbstractAlgebra.gens(G::Generic.SymmetricGroup)
+    a = one(G)
+    b = one(G)
+    circshift!(b.d, a.d, -1)
+    a.d[1], a.d[2] = 2, 1
+    return [a, b]
+end
